refactor: rename root scripts folder from "script" to "u1_image_base" to prevent namespace conflicts
