+++ skills/u1-image-base/u1_image_base/llm/chat_completions_adapter.py
