+++ skills/u1-image-base/u1_image_base/llm/llm_adapter.py
