+++ skills/u1-image-base/u1_image_base/vlm/chat_completions_adapter.py
