+++ skills/u1-image-base/u1_image_base/vlm/vlm_adapter.py
